ecocycle-hub/requirements.txt
Flask==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.1
werkzeug==3.1.8
Flask-SQLAlchemy
Flask-Login
Flask-WTF
